+++ EntraAuth/changelog.md
@@ -1,6 +1,14 @@
 ﻿# Changelog
 
-## 1.0.6
+## 1.1.11 (2024-05-21)
+
++ New: Service configurations - Added configurations for Azure & AzureKeyVault.
++ Upd: Connect-EntraService - Added support for direct Key Vault integration.
++ Upd: Service configurations - Added capability to require additional parameters that modify the base Service Url.
++ Fix: Token Renewal - bad parameter ServiceUrl.
++ Fix: Asset-EntraConnection - bad error message when assertion fails.
+
+## 1.0.6 (2024-05-15)
 
 + Upd: Invoke-EntraRequest - added -NoPaging parameter to support disabling paging.
 + Upd: Invoke-EntraRequest - added -Raw parameter to support returning unprocessed results.